Fix build with OpenSSL 1.1 due to RSA being an opaque struct
RSA is an opaque struct in OpenSSL 1.1. New getter functions must be
used to access the key components. The functions were not present in
OpenSSL 1.0, so add a compat header with the implementation of the
needed functions as suggested by the OpenSSL wiki [1] in order to allow
building tpm-tools with any version of OpenSSL.

[1] https://wiki.openssl.org/index.php/1.1_API_Changes

Signed-off-by: Debora Velarde Babb <debora@linux.ibm.com>